# Cache task_vars to speed up IncludedFile.process_include_results