##### SUMMARY
Cache task_vars to speed up IncludedFile.process_include_results

`process_include_results` can be slow.  When profiling a loop on an `include_tasks` I noticed that the majority of the time spent were calls to `get_vars`.  This PR aims to cache them to improve performance.


##### ISSUE TYPE
 - Bugfix Pull Request

##### COMPONENT NAME
lib/ansible/playbook/included_file.py

##### ANSIBLE VERSION
<!--- Paste verbatim output from "ansible --version" between quotes below -->
```
devel
```


##### ADDITIONAL INFORMATION

Some perf stats from cProfile:

before:
```
TASK [include_tasks] *************************************************************************************************************************************************************
         2195754 function calls (2184229 primitive calls) in 2.094 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.010    0.010    2.098    2.098 included_file.py:70(process_include_results)
      461    0.027    0.000    1.955    0.004 manager.py:154(get_vars)
    10603    0.077    0.000    0.946    0.000 loader.py:423(all)
      461    0.009    0.000    0.778    0.002 manager.py:430(_get_magic_variables)
```



after:
```
TASK [include_tasks] *************************************************************************************************************************************************************
         156118 function calls (156093 primitive calls) in 0.112 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.005    0.005    0.112    0.112 included_file.py:70(process_include_results)
      461    0.020    0.000    0.056    0.000 {method 'index' of 'list' objects}
   106030    0.035    0.000    0.035    0.000 included_file.py:64(__eq__)
      461    0.002    0.000    0.022    0.000 dataloader.py:177(path_dwim)
```

The above done using:

```
---
- hosts: localhost
  gather_facts: false
  tasks:
    - include_tasks: include_me.yml
      with_sequence: start=0 end=460
```