@sivel 2018-04-19
Timing with 100 hosts and the 460 loops:

before:
```
real	3m24.319s
```

after:
```
real	0m51.863s
```