@sivel 2018-04-19
I just added another performance improvement to use a dict to track `included_files` instead of a list.

before:
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.545    0.545   12.286   12.286 included_file.py:54(process_include_results)
    46100    1.997    0.000    5.540    0.000 {method 'index' of 'list' objects}
 10648639    3.542    0.000    3.542    0.000 included_file.py:48(__eq__)
```

after:
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.549    0.549    6.933    6.933 included_file.py:70(process_include_results)
    46100    0.167    0.000    2.124    0.000 dataloader.py:177(path_dwim)
    46100    0.179    0.000    1.640    0.000 path.py:30(unfrackpath)
```

On the 100 hosts example + 460 loops, this cuts off over 5 additional seconds.

Per a conversation @jimi-c and I had, the fact that dicts may be unordered is of no consequence here, as the particular ordering wasn't guaranteed to be consistent to begin with.

As an apples to apples comparison, the 100 hosts + 460 item loop, without this PR ansible spends `229.646s` in `process_include_results`.  So this PR reduces that time by `222.713s`.