@sivel 2018-04-20
I've backed out change to track included files using a dict.  The performance change wasn't worth the added complexity at this time.